+++ gsclient/storage.go
@@ -16,7 +16,7 @@ type StorageProperties struct {
 	Status          string  `json:"status"`
 	CreateTime      string  `json:"create_time"`
 	ChangeTime      string  `json:"change_time"`
-	CurrentPrice    float32 `json:"current_price"`
+	CurrentPrice    float64 `json:"current_price"`
 	LocationName    string  `json:"location_name"`
 	LocationCountry string  `json:"location_country"`
 	LocationIata    string  `json:"location_iata"`
